feat: 전화번호 저장 메뉴주문시에도 추가

+++ src/screen/order/OrderPhoneScreen.tsx
@@ -20,8 +20,10 @@ export default function OrderPhoneScreen() {
         const { data } = r;
         if (data?.length !== 0) {
           dispatch({ type: 'SET_VIEW', view: OrderView.CHOOSE_ORDER });
-          if (window.confirm(`${phone} 번호를 저장하시겠습까?`)) {
-            localStorage.setItem('phone', phone);
+          if (!localStorage.getItem('phone')) {
+            if (window.confirm(`${phone} 번호를 저장하시겠습까?`)) {
+              localStorage.setItem('phone', phone);
+            }
           }
         } else {
           alert('주문이 없습니다');
